(test) Fix Test Class for LIC14 and Fix Testing Convention for LIC14
